refactor: move dto files to be separated by req or res
