--- src/main/kotlin/com/mansaeng/diridibackend/dto/AuthResponse.kt
@@ -1,3 +1,0 @@
-package com.mansaeng.diridibackend.dto
-
-data class AuthResponse(val token: String)
