+++ src/main/kotlin/com/mansaeng/diridibackend/dto/request/AuthRequest.kt
@@ -1,4 +1,4 @@
-package com.mansaeng.diridibackend.dto
+package com.mansaeng.diridibackend.dto.request
 
 data class AuthRequest(val username: String, val password: String) {
