+++ src/main/kotlin/com/mansaeng/diridibackend/dto/request/CreateEpisodeRequest.kt
@@ -1,4 +1,4 @@
-package com.mansaeng.diridibackend.dto
+package com.mansaeng.diridibackend.dto.request
 
 import java.time.LocalDateTime
