+++ src/main/kotlin/com/mansaeng/diridibackend/dto/response/AuthResponse.kt
@@ -1,0 +1,3 @@
+package com.mansaeng.diridibackend.dto.response
+
+data class AuthResponse(val token: String)
